+++ app/core/Engine/controllers/transaction-controller/transaction-controller-init.ts
@@ -3,6 +3,9 @@ import {
   TransactionType,
   type TransactionControllerMessenger,
   type TransactionMeta,
+  type PublishBatchHookRequest,
+  type PublishBatchHookTransaction,
+  type PublishBatchHookResult,
 } from '@metamask/transaction-controller';
 import { SmartTransactionStatuses } from '@metamask/smart-transactions-controller/dist/types';
 import { hasProperty, Hex } from '@metamask/utils';
@@ -18,8 +21,10 @@ import Logger from '../../../../util/Logger';
 import { getGlobalChainId as getGlobalChainIdSelector } from '../../../../util/networks/global-network';
 import {
   submitSmartTransactionHook,
+  submitBatchSmartTransactionHook,
   type SubmitSmartTransactionRequest,
 } from '../../../../util/smart-transactions/smart-publish-hook';
+import { getTransactionById } from '../../../../util/transactions';
 import type { RootState } from '../../../../reducers';
 import { TransactionControllerInitMessenger } from '../../messengers/transaction-controller-messenger';
 import type {
@@ -81,7 +86,8 @@ export const TransactionControllerInit: ControllerInitFunction<
           networkController.getNetworkClientRegistry(...args),
         getNetworkState: () => networkController.state,
         hooks: {
-          publish: (transactionMeta: TransactionMeta) =>
+          // @ts-expect-error - TransactionController actually sends a signedTx as a second argument, but its type doesn't reflect that.
+          publish: (transactionMeta: TransactionMeta, signedTransactionInHex: Hex) =>
             publishHook({
               transactionMeta,
               getState,
@@ -89,6 +95,17 @@ export const TransactionControllerInit: ControllerInitFunction<
               smartTransactionsController,
               approvalController,
               initMessenger,
+              signedTransactionInHex,
+            }),
+          publishBatch: async (_request: PublishBatchHookRequest) =>
+            await publishBatchSmartTransactionHook({
+              transactionController,
+              smartTransactionsController,
+              initMessenger,
+              getState,
+              approvalController,
+              transactions:
+                _request.transactions as PublishBatchHookTransaction[],
             }),
         },
         incomingTransactions: {
@@ -132,6 +149,7 @@ function publishHook({
   smartTransactionsController,
   approvalController,
   initMessenger,
+  signedTransactionInHex
 }: {
   transactionMeta: TransactionMeta;
   getState: () => RootState;
@@ -139,9 +157,11 @@ function publishHook({
   smartTransactionsController: SmartTransactionsController;
   approvalController: ApprovalController;
   initMessenger: TransactionControllerInitMessenger;
+  signedTransactionInHex: Hex;
 }): Promise<{ transactionHash: string }> {
   const state = getState();
-  const shouldUseSmartTransaction = selectShouldUseSmartTransaction(state, transactionMeta.chainId);
+  const { shouldUseSmartTransaction, featureFlags } =
+    getSmartTransactionCommonParams(state, transactionMeta.chainId);
 
   // @ts-expect-error - TransactionController expects transactionHash to be defined but submitSmartTransactionHook could return undefined
   return submitSmartTransactionHook({
@@ -152,7 +172,69 @@ function publishHook({
     approvalController,
     controllerMessenger:
       initMessenger as unknown as SubmitSmartTransactionRequest['controllerMessenger'],
-    featureFlags: selectSwapsChainFeatureFlags(state),
+    featureFlags,
+    signedTransactionInHex,
+  });
+}
+
+function getSmartTransactionCommonParams(state: RootState, chainId?: Hex) {
+  const shouldUseSmartTransaction = selectShouldUseSmartTransaction(state, chainId);
+  const featureFlags = selectSwapsChainFeatureFlags(state, chainId);
+
+  return {
+    shouldUseSmartTransaction,
+    featureFlags,
+  };
+}
+
+function publishBatchSmartTransactionHook({
+  transactionController,
+  smartTransactionsController,
+  initMessenger,
+  getState,
+  approvalController,
+  transactions,
+}: {
+  transactionController: TransactionController;
+  smartTransactionsController: SmartTransactionsController;
+  initMessenger: TransactionControllerInitMessenger;
+  getState: () => RootState;
+  approvalController: ApprovalController;
+  transactions: PublishBatchHookTransaction[];
+}): Promise<PublishBatchHookResult> {
+  // Get transactionMeta based on the last transaction ID
+  const lastTransaction = transactions[transactions.length - 1];
+  const transactionMeta = getTransactionById(
+    lastTransaction.id ?? '',
+    transactionController,
+  );
+  const state = getState();
+
+  if (!transactionMeta) {
+    throw new Error(
+      `publishBatchSmartTransactionHook: Could not find transaction with id ${lastTransaction.id}`,
+    );
+  }
+
+  const { shouldUseSmartTransaction, featureFlags } =
+    getSmartTransactionCommonParams(state, transactionMeta.chainId);
+
+  if (!shouldUseSmartTransaction) {
+    throw new Error(
+      'publishBatchSmartTransactionHook: Smart Transaction is required for batch submissions',
+    );
+  }
+
+  return submitBatchSmartTransactionHook({
+    transactions,
+    transactionController,
+    smartTransactionsController,
+    controllerMessenger:
+      initMessenger as unknown as SubmitSmartTransactionRequest['controllerMessenger'],
+    shouldUseSmartTransaction,
+    approvalController,
+    featureFlags,
+    transactionMeta
   });
 }