+++ e2e/selectors/wallet/WalletView.selectors.js
@@ -82,6 +82,9 @@ export const WalletViewSelectorsIDs = {
   CAROUSEL_SIXTH_SLIDE: 'carousel-sixth-slide',
   CAROUSEL_SIXTH_SLIDE_TITLE: 'carousel-sixth-slide-title',
   CAROUSEL_SIXTH_SLIDE_CLOSE_BUTTON: 'carousel-sixth-slide-close-button',
+  CAROUSEL_SEVENTH_SLIDE: 'carousel-seventh-slide',
+  CAROUSEL_SEVENTH_SLIDE_TITLE: 'carousel-seventh-slide-title',
+  CAROUSEL_SEVENTH_SLIDE_CLOSE_BUTTON: 'carousel-seventh-slide-close-button',
   CAROUSEL_PROGRESS_DOTS: 'progress-dots',
   CAROUSEL_SLIDE: 'carousel-slide',
 };